fix(material/core): avoid running sanity checks on some test environments
Fixes that the sanity checks were running on some testing environments like Jest.

These changes also move the test environment check into a centralized place so that it's easier to keep up to date.

Fixes #23365.
