+++ src/cdk/platform/public-api.ts
@@ -12,3 +12,4 @@ export * from './features/input-types';
 export * from './features/passive-listeners';
 export * from './features/scrolling';
 export * from './features/shadow-dom';
+export * from './features/test-environment';
